release_package/pipeline/requirements.txt
# SPECTRA sanitized parsing pipeline dependencies
# Tested with Python 3.10+

# Document parsing
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0

# RDF / SHACL validation
rdflib>=7.0.0
pyshacl>=0.25.0

# Neo4j bulk load
neo4j>=5.18.0

# 3GPP portal scraping (Stage 01)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Misc
python-dotenv>=1.0.0
